moved sort to selector, it was causing view to refresh/resort markets. renamed index.js file to more descriptive name so it can be found easily. updated tests. Reporting Reporting page list of upcoming [ch8648]
