+++ src/modules/reporting/components/reporting-report-markets/reporting-report-markets.jsx
@@ -25,8 +25,7 @@ export const ReportSection = ({
   if (items.length === 0) {
     theChildren = <NoMarketsFound message={emptyMessage} key={title} />
   } else {
-    theChildren = items.sort((a, b) => a.endDate.timestamp - b.endDate.timestamp)
-      .map(item => (<MarketPreview key={item.id} buttonText={buttonText} {...item} />))
+    theChildren = items.map(item => (<MarketPreview key={item.id} buttonText={buttonText} {...item} />))
   }
 
   return (
