+++ src/modules/reporting/containers/reporting-report-markets.js
@@ -3,7 +3,7 @@ import { withRouter } from 'react-router-dom'
 
 import ReportingReportMarkets from 'modules/reporting/components/reporting-report-markets/reporting-report-markets'
 import { loadReporting } from 'src/modules/reporting/actions/load-reporting'
-import { selectMarketsToReport } from 'src/modules/reporting/selectors'
+import { selectMarketsToReport } from 'src/modules/reporting/selectors/select-markets-to-report'
 
 const mapStateToProps = state => ({
   isLogged: state.isLogged,
