+++ src/modules/reporting/containers/reporting-resolved.js
@@ -4,7 +4,7 @@ import { withRouter } from 'react-router-dom'
 import ReportingResolved from 'modules/reporting/components/reporting-resolved/reporting-resolved'
 // import { loadResolved } from 'modules/reporting/actions/load-resolved'
 import { loadReporting } from 'src/modules/reporting/actions/load-reporting'
-import { selectMarketsToReport } from 'src/modules/reporting/selectors'
+import { selectMarketsToReport } from 'src/modules/reporting/selectors/select-markets-to-report'
 import { toggleFavorite } from 'modules/markets/actions/update-favorites'
 import { loadMarketsInfo } from 'modules/markets/actions/load-markets-info'
