--- src/modules/reporting/selectors/index.js
@@ -1,13 +1,0 @@
-import { createSelector } from 'reselect'
-import { selectMarketReportState } from 'src/select-state'
-import { selectMarket } from 'src/modules/market/selectors/market'
-
-export const selectMarketsToReport = createSelector(
-  selectMarketReportState,
-  reportData => ({
-    designated: reportData.designated.map(selectMarket),
-    open: reportData.open.map(selectMarket),
-    upcoming: reportData.upcoming.map(selectMarket),
-    resolved: reportData.resolved.map(selectMarket),
-  }),
-)
